Update HourlyStat to use Float share_count

+++ app/models/hourly_stat.rb
@@ -6,5 +6,5 @@ class HourlyStat
   field :date,        type: String
   field :hour,        type: Integer
   field :order_count, type: Integer
-  field :share_count, type: Integer
+  field :share_count, type: Float
 end
